+++ src/App.jsx
@@ -1,54 +1,50 @@
-/*  eslint react-hooks/exhaustive-deps: off  */
-// 上記文でeslintのエラー検知をオフにしてくれる
-import React, { useEffect, useState } from 'react'
-import ColorfulMessage from './components/ColorfulMessage';
+import React, { useState } from 'react'
+import "./style.css";
 
-// ファイルの中身がコンポーネントだよと明示する場合はファイルの拡張子はjsxにする
-const App = () => {
-  const [num, setNum] = useState(0);
-  const [faceShowFalg, setFaceShowFlag] = useState(false);
-  const onClickCountButton = () => {
-    setNum(num + 1);
-  };
-
-  const onClickSwitchShowFlag = () => {
-    setFaceShowFlag(!faceShowFalg);
-  }
-
-  useEffect(() => {
-    if (num > 0) {
-      if (num % 3 === 0) {
-        faceShowFalg || setFaceShowFlag(true) //  ||は左側の式がfalseだったら右側の処理を走らせる
-      } else {
-        faceShowFalg && setFaceShowFlag(false) //  &&は左側の式がtrueだったら右側の処理を走らせる
-      }
-    }
-    //eslintオフの記述は一文のみへの適用も可能
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [num]);// 変数をセットすると、セットした変数が変化したときにのみ、処理が走る
-
-
-
-  const contentStyle = {
-    color: 'blue',
-    fontSize: '18px'
-  };
-
+export const App = () => {
+  const [inCompleteTodos, setInCompleteTodos] = useState([
+    "あああ",
+    "いいい"
+  ]);
+  const [completeTodos, setCompleteTodos] = useState([
+    "ううう",
+    "えええ"
+  ]);
   return (
     <>
-      <h1 style={{ color: 'red' }}>こんにちは!!</h1>
-      <p style={contentStyle}>ハロー！！</p>
-      <ColorfulMessage color="blue"> お元気ですか？</ColorfulMessage>
-      <ColorfulMessage color="pink" message="元気です" />
+      <div className="add_todo_area">
+        <input id="add-text" placeholder="TODOを入力" type="text"></input>
+        <button id="add-button">追加</button>
+      </div>
 
-      <button onClick={onClickCountButton}>カウントアップ</button>
-      <br />
-      <button onClick={onClickSwitchShowFlag}>on/off</button>
-      <p>{num}</p>
-      {faceShowFalg && <p>(#^.^#)</p>}
+      <div className="incomplete_area content_area">
+        <h2>未完了のTODO</h2>
+        <ul id="inComplete_list">
+          {inCompleteTodos.map((todo) => {
+            return (
+              <li key={todo} className="list_item">
+                <p className="list_name">{todo}</p>
+                <button>完了</button>
+                <button>削除</button>
+              </li>
+            )
+          })}
+        </ul>
+      </div>
+
+      <div className="complete_area content_area2">
+        <h2>完了したTODO</h2>
+        <ul className="complete_list">
+          {completeTodos.map((todo) => {
+            return (
+              <li key={todo} className="list_item">
+                <p className="list_name">{todo}</p>
+                <button>戻す</button>
+              </li>
+            )
+          })}
+        </ul>
+      </div>
     </>
   )
-
-}
-
-export default App;
+};
