+++ src/style.css
@@ -1,0 +1,82 @@
+* {
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  padding: 100px;
+}
+
+h2 {
+  text-align: center;
+  margin-bottom: 50px;
+}
+
+.add_todo_area {
+  background-color: aqua;
+  padding: 30px;
+  display: flex;
+  align-items: center;
+  border-radius: 20px;
+  margin-bottom: 60px;
+}
+
+.add_todo_area input {
+  width: 50%;
+  height: 50px;
+  border-radius: 20px;
+}
+
+.add_todo_area input::placeholder {
+  padding-left: 20px;
+  font-size: 20px;
+}
+
+.add_todo_area button {
+  height: 50px;
+  width: 150px;
+  font-size: 20px;
+  margin-left: 10px;
+  border-radius: 20px;
+}
+
+
+.content_area {
+  background-color: rgb(100, 237, 109);
+  padding: 60px;
+  border-radius: 20px;
+  margin-bottom: 60px;
+}
+
+.content_area ul li,
+.content_area2 ul li {
+  margin-bottom: 40px;
+}
+
+.list_item {
+  display: flex;
+  align-items: center;
+  font-size: 20px;
+}
+
+.content_area button {
+  font-size: 20px;
+  height: 50px;
+  width: 150px;
+  margin-left: 20px;
+  border-radius: 20px;
+}
+
+.content_area2 {
+  background-color: rgb(219, 230, 14);
+  padding: 60px;
+  border-radius: 20px;
+}
+
+.content_area2 button {
+  font-size: 20px;
+  height: 50px;
+  width: 150px;
+  margin-left: 20px;
+  border-radius: 20px;
+}
